aetheranime/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
pillow==11.0.0
psycopg2-binary==2.9.10
social-auth-app-django==5.4.2
social-auth-core==4.5.4
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
aiohttp==3.11.10
drf-spectacular==0.25.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-django==4.9.0
drf-yasg
django-extensions
gunicorn
locust
django-allauth
wagtail
django-taggit